22 de julio

+++ WebApiCaracterizacion/DataGanaderia/PromedioEtniaGNRepository.cs
@@ -68,7 +68,7 @@ namespace WebApiCaracterizacion.DataGanaderia
                 dato = (string)reader["dato"],
                 cantidad = (int)reader["cantidad"],
                 porcentaje = (double)reader["porcentaje"],
-                orden = (int)reader["orden"],
+                orden = (double)reader["orden"],
                 color = (string)reader["color"],
                 nombre_campana = (string)reader["nombre_campana"]
             };
@@ -80,7 +80,7 @@ namespace WebApiCaracterizacion.DataGanaderia
                 dato = (string)reader["dato"],
                 cantidad = (int)reader["cantidad"],
                 porcentaje = (double)reader["porcentaje"],
-                orden = (int)reader["orden"],
+                orden = (double)reader["orden"],
                 color = (string)reader["color"],
                 nombre_campana = (string)reader["nombre_campana"]
             };
@@ -94,7 +94,7 @@ namespace WebApiCaracterizacion.DataGanaderia
                 dato = (string)reader["dato"],
                 cantidad = (int)reader["cantidad"],
                 porcentaje = (double)reader["porcentaje"],
-                orden = (int)reader["orden"],
+                orden = (double)reader["orden"],
                 color = (string)reader["color"],
                 nombre_campana = (string)reader["nombre_campana"]
             };
@@ -108,7 +108,7 @@ namespace WebApiCaracterizacion.DataGanaderia
                 dato = (string)reader["dato"],
                 cantidad = (int)reader["cantidad"],
                 porcentaje = (double)reader["porcentaje"],
-                orden = (int)reader["orden"],
+                orden = (double)reader["orden"],
                 color = (string)reader["color"],
                 nombre_campana = (string)reader["nombre_campana"]
             };
