+++ WebApiCaracterizacion/DataMineria/PromedioVariacionORRepository.cs
@@ -124,7 +124,7 @@ namespace WebApiCaracterizacion.DataMineria
                 dato = (string)reader["dato"],
                 cantidad = (int)reader["cantidad"],
                 porcentaje = (double)reader["porcentaje"],
-                orden = (int)reader["orden"],
+                orden = (double)reader["orden"],
                 color = (string)reader["color"],
                 nombre_campana = (string)reader["nombre_campana"]
 
@@ -140,7 +140,7 @@ namespace WebApiCaracterizacion.DataMineria
                 dato = (string)reader["dato"],
                 cantidad = (int)reader["cantidad"],
                 porcentaje = (double)reader["porcentaje"],
-                orden = (int)reader["orden"],
+                orden = (double)reader["orden"],
                 color = (string)reader["color"],
                 nombre_campana = (string)reader["nombre_campana"]
             };
@@ -155,7 +155,7 @@ namespace WebApiCaracterizacion.DataMineria
                 dato = (string)reader["dato"],
                 cantidad = (int)reader["cantidad"],
                 porcentaje = (double)reader["porcentaje"],
-                orden = (int)reader["orden"],
+                orden = (double)reader["orden"],
                 color = (string)reader["color"],
                 nombre_campana = (string)reader["nombre_campana"]
             };
@@ -169,7 +169,7 @@ namespace WebApiCaracterizacion.DataMineria
                 dato = (string)reader["dato"],
                 cantidad = (int)reader["cantidad"],
                 porcentaje = (double)reader["porcentaje"],
-                orden = (int)reader["orden"],
+                orden = (double)reader["orden"],
                 color = (string)reader["color"],
                 nombre_campana = (string)reader["nombre_campana"]
             };
