+++ WebApiCaracterizacion/DataTransporte/PromedioOriginarioTFRepository.cs
@@ -59,7 +59,7 @@ namespace WebApiCaracterizacion.Data
                 dato = (string)reader["dato"],
                 cantidad = (int)reader["cantidad"],
                 porcentaje = (double)reader["porcentaje"],
-                orden = (int)reader["orden"],
+                orden = (double)reader["orden"],
                 color = (string)reader["color"],
                 nombre_campana = (string)reader["nombre_campana"]
 
@@ -72,7 +72,7 @@ namespace WebApiCaracterizacion.Data
                 dato = (string)reader["dato"],
                 cantidad = (int)reader["cantidad"],
                 porcentaje = (double)reader["porcentaje"],
-                orden = (int)reader["orden"],
+                orden = (double)reader["orden"],
                 color = (string)reader["color"],
                 nombre_campana = (string)reader["nombre_campana"]
             };
