+++ WebApiCaracterizacion/ModelsMineria/PromediosVariacionOR.cs
@@ -7,7 +7,7 @@
         public string dato { get; set; }
         public int cantidad { get; set; }
         public double porcentaje { get; set; }
-        public int orden { get; set; }
+        public double orden { get; set; }
         public string color { get; set; }
         public string nombre_campana { get; set; }
     }
